pyproject.toml
[project]
name = "aimi"
version = "0.1.0"
description = "Aimi is a personal AI coach that tracks goals, plans tasks, and keeps conversational memory across sessions."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "openai>=1.108.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
    "python-json-logger>=2.0",
    "uvicorn>=0.30",
    "websockets>=12.0",
    "redis>=5.2",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "greenlet>=3.0",
    "pyjwt>=2.9",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.28",
    "pre-commit>=3.7",
    "ruff>=0.5",
    "mypy>=1.10",
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "fakeredis>=2.23",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
exclude = ["alembic/", "scripts/generate_tokens.py"]

[tool.ruff.format]
exclude = ["alembic/", "scripts/generate_tokens.py"]

[tool.mypy]
python_version = "3.13"

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true
